[DEC-2081] Audit app/simulation_state.go
All these functions are only used in tests and we are blocked from using the original functions since we forked them to address DEC-1859.
